+++ docs/source/reference/api/metakb.database.rst
@@ -1,0 +1,8 @@
+﻿metakb.database
+===============
+
+.. automodule:: metakb.database
+   :members:
+   :undoc-members:
+   :special-members: __init__
+   :exclude-members: model_fields, model_config, model_computed_fields
