+++ docs/source/reference/api/metakb.load_data.rst
@@ -1,0 +1,8 @@
+﻿metakb.load_data
+================
+
+.. automodule:: metakb.load_data
+   :members:
+   :undoc-members:
+   :special-members: __init__
+   :exclude-members: model_fields, model_config, model_computed_fields
